Patch for building MRI trunk and JRuby.
This patch will run autoconf if it sees a configure.in file in the
source root but there's no configure script. This allows you to
point ry at the MRI git repo and it will run autoconf first and
then build it as usual. There's also a patch to be able to pass
args to the Rubinius installer script. Also in this patch is new
code path to build JRuby.