Fix incorrect exit status of the fullpath command.
'ry fullpath myruby' does not return the correct exit
status when 'myruby' is not known to ry. This is because
'fullpath' calls 'binpath' in a separate subshell due to the use
of pipes(piping to tr). This commit fixes this.

Additionally, 'fullpath' now adds a newline at the end to make
the full path looks nicer. Omitting the newline in the first place
was not necessary since using $(ry fullpath myruby) will automatically
strip the trailing newline.